requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pika==1.3.1
pika-stubs==0.1.3
psycopg2==2.9.5
python-dotenv==1.0.0
requests==2.28.2
schedule==1.1.0
urllib3==1.26.14
Werkzeug==2.2.3
